services/multi-criteria-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
pydantic==2.7.1
numpy==1.26.4